feat: Refactored AppAccountSelectorView to be usable from different views
